Fixed GCC Compilation Issues
+ Got the new scanning functionality working on Mint Linux with GCC 7.2.

+++ source/Utilities/viewFrustum.hpp
@@ -97,7 +97,7 @@ namespace FrustumUtilities
       std::vector<std::pair<float, float>> cascadeDistances;
       for (auto index{ 1.0 }; index < cascadeCount; ++index)
       {
-         const auto cascade = nearPlane * std::powf(planeRatio, index / cascadeCount);
+         const auto cascade = nearPlane * std::pow(planeRatio, index / cascadeCount);
          cascadeDistances.emplace_back(previousCascadeStart, cascade);
          previousCascadeStart = cascade;
       }
